requirements.txt
appnope==0.1.4
asttokens==3.0.0
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.0
fonttools==4.57.0
imbalanced-learn==0.13.0
imblearn==0.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numba==0.61.2
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.4.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shap==0.47.2
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
xgboost==3.0.1
